+++ public/dev_client/config.php
@@ -3,5 +3,5 @@ $port = 8010;
 $client_click_name = "client_click_dev.php";
 $client_functions_name = "client_functions_dev.php";
 $client_network_name = "client_network_dev.php";
-$environment = "production";
+$environment = "development";
 ?>